RoverInterface/requirements.txt
nicegui
opencv-python-headless
requests
pyserial
ultralytics>=8.4.0
mlx-vlm>=0.3.9
pillow>=10.0.0
numpy>=1.24.0